Updated endecrypt function to return int

+++ rsa.h
@@ -94,7 +94,7 @@ long coprime(long x) {
 }
 
 // Given an integer representing an ASCII value, encrypt it via RSA
-long endecrypt(long msg, long key, long c) {
+int endecrypt(int msg, long key, long c) {
   return modulo(msg, key, c);
 }
